# Use swizzling instead of padding