- Remove performance report on README for maintenance simplicity

- Peak TFLOPS 1503 -> 1519 TFLOPS, the original 1520 is on another "good" GPU
  - `(m= 4096, n=24576, k= 1536)`: 1212 -> 1270 TFLOPS
  - `(m= 4096, n=32768, k=  512)`: 775 -> 836 TFLOPS
  - `(m= 4096, n= 7168, k=16384)`: 1503 -> 1519 TFLOPS
  - `(m= 4096, n= 4096, k= 7168)`: 1410 -> 1439 TFLOPS
  - `(m= 4096, n= 7168, k= 2048)`: 1168 -> 1215 TFLOPS